tests/requirements.txt
numpy
google_genai
pillow